refactor: change route from organisation to organisations (plural) and update the links
