+++ frontend/config/menuItems.ts
@@ -13,7 +13,7 @@ export const menuItems:MenuItemType[] = [
   // {path:"/", label:"Home"},
   {path:'/software', label:'Software'},
   {path: '/projects', label: 'Projects'},
-  {path: '/organisation', label: 'Organisations'},
+  {path: '/organisations', label: 'Organisations'},
   // will be done later
   // {path:"/about", label:"About"},
 ]
